Add option to mesure webpack speed
