+++ examples/getstarted/package.json
@@ -1,7 +1,7 @@
 {
   "name": "getstarted",
   "private": true,
-  "version": "4.2.0-beta.0",
+  "version": "4.1.4",
   "description": "A Strapi application.",
   "scripts": {
     "develop": "strapi develop",
